Rewrite for dependency tracking
This significantly reworks the datastructures used to encode the
historical stdlib data, so that we can include dependencies and weak
dependencies for all stdlib versions.  This will be used by Pkg to
properly resolve transitive dependencies of stdlibs.

This also changes `HistoricalStdlibVersions` to register the data
structures with `Pkg` on load.  This simplifies things greatly, and also
allows us to maintain a single location that knows how to deal with
multiple different `Pkg` versions, as only Julia v1.12+ will know how to
use the newer datastructures here.