compress the version map a bit
we use a compression where for a given stdlib with given dependencies we associate that with a julia range. Within that julia range we also store how the version of the stdlib itself has evolved.

The compression and uncompression code were written by an LLM after my instruction of how it should be compressed.